layoutlm/requirements.txt
torch>=1.10.0
transformers>=4.17.0
pdf2image>=1.16.0
Pillow>=9.0.0
matplotlib>=3.5.0
opencv-python>=4.5.5
tqdm>=4.62.0
scikit-learn>=1.0.0
pandas>=1.3.5
paddlepaddle>=2.2.0
paddleocr>=2.5.0
numpy>=1.21.0 